Description of CBuchan_A1.java
The main class called Health Professional is the blueprint parent class of the two child classes called GP and Nurse. The Appointment class was created to handle functions such as creating a new appointment, cancelling an existing appointment and printing all appointments. This Appointment class is technically also a child class of the Health Professional. The function of this program is handled within the public main method, which creates objects of those mentioned classes and calls upon existing methods from those object classes, like .printMethod(). The purpose was to create an appointment program that handles patients' data and information, parsing into arrays and instance variables to represent an actual database and how a real appointment system for a health system may be handled. This was uploaded as the end result. My goal in future endeavours is to upload in intervals to demonstrate my future programs' progression.